Uses uptime_lib for retrieving uptime
uptime_lib supports a superset of what systemstat does right now, so
this change fully delegates that logic to uptime_lib.

The target_os list in src/lib.rs may not be necessary but it at least
will break a build on any OSes that systemstat decides to support that
uptime_lib does not yet support as of uptime_lib 0.1.0.