Fix for Windows < 10 and Removable/CDROM drives
On Windows 10, CDROM and Removable drives (Floppy, USB,
etc) seem to be ignored unless they are currently active (meaning that
there is media mounted to these drives). In previous versions of
Windows, the OS seems to consider those drives valid and therefore
attempts to get volume information from them and emits the `OS
error 21 (device is not ready)` error. To fix this, on error of the
`GetVolumeInformationW` system call, the function calls
`GetDriveTypeW` and checks whether the drive is removable or a
CDROM. If so, set all disk space information to 0 and carry on.