Added Serialize and Deserialize traits (#85)
* Added, and updated dependencies for serde capablilties

serde is the most used crate for serializing and de-serializing data (structs, enums, etc). Added dependencies for that

* Added Serialize and Deserialize traits 

Added Serialize and Deserialize traits to public structures. This allows user to use serde to serialize and deserialize them

* Added serde library

the serde library is the most used serializing and de-serializing library. This allows for users to serialize and de-serialize structures, enums, etc.

* Added serialize and deserialize for a PlatformCpuLoad

Forgot to added it for PlatformCpuLoad

* Added serde-feature-hack

added serde-feature-hack to add serde feautres

* added external crate serde

Added the external crate serde for macros

* Added serialize and deserialize as serde feature only

Added serialize and deserialize to be active only if the serde feature is activated